requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
idna==3.6
newsapi-python==0.2.7
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
